library_frontend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.1
django-ninja==1.3.0
ninja-put-patch-file-upload-middleware==0.1.4
orjson==3.10.15
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
